requirements.txt
# Adamus Core Dependencies
# Python 3.11+ required

# AI APIs
anthropic>=0.18.0
ollama>=0.1.0

# Database
aiosqlite>=0.19.0

# Security
python-dotenv>=1.0.0
cryptography>=42.0.0

# Utilities
pyyaml>=6.0
aiofiles>=23.0.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Web (for War Room later)
flask>=3.0.0
